+++ src/modules/esm.mjs
@@ -1,0 +1,37 @@
+import path, { dirname } from 'path';
+import fs from 'fs';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { fileURLToPath } from 'url';
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(fileURLToPath(import.meta.url));
+import './files/c.js';
+
+const random = Math.random();
+
+const pathToObject = random > 0.5 ? './files/a.json' : './files/b.json';
+
+const rawdata = fs.readFileSync(path.join(__dirname, pathToObject));
+const unknownObject = JSON.parse(rawdata);
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { myServer, unknownObject };
